refactor(vm): split vm_set_ops.rs into cohesive submodules (§7-8)
`vm_set_ops.rs` (866 lines) exceeded the 500-line module convention. Split into
two cohesive files, byte-identically (function multiset verified unchanged;
methods reached via `self`):

- `vm_set_ops.rs` (460): membership/coercion helpers (range/hash/set contains,
  value->set/bag/mix key coercion, quant-hash subset checks) and the core set
  ops `(elem)` / `(cont)` / `(|)`.
- `vm_set_arith_ops.rs` (412): the arithmetic/comparison set ops
  (`(&)`/`(.)`/`(-)`/`(^)`/`(<=)`/`(<)`/...), Bag/Mix item coercion, and the
  junction ops (`any`/`all`/`one`/n-ary).

Three A-group helpers called from the arithmetic ops (`quant_hash_subset`,
`quant_hash_strict_subset`, `lazy_list_error`) were bumped to `pub(super)`.
Both files are now under 500 lines.
